Before:
pip3 install .  43.18s user 2.96s system 99% cpu 46.516 total

After:
pip3 install .  62.00s user 3.36s system 358% cpu 18.257 total

Similar speedups for luarocks.