@jgehring 2017-06-15
Actually, we could just use cmake? Right now it's mildly annoying that changing a single cpp file of the Python wrapper requires recompiling everything...