@ebetica 2017-06-15
The right way is to link the library for generated by the cmake I think.